moved some stuff to a util file for the crate
